requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==45.0.5
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.6
python-multipart==0.0.6
greenlet==3.2.3
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
